refactor: refactoring the app
It was deleted all descriptive message from de css files and js file and only the necessary one was left, also was created a new js file named scroll.js for make the code more dry
