+++ src/css/footer-tkp.css
@@ -11,10 +11,8 @@ footer {
 	position: absolute;
 	width: 100vw;
 	height: 146px;
-	/* height: auto; */
 	padding: 30px 0 0 0;
 	bottom: 0;
-	/* z-index: -1; */
 	color: white;
 	background: var(--blue);
 }
@@ -59,7 +57,6 @@ footer {
 	flex-direction: column;
 
 	height: 88px;
-	/* margin: 1rem 0 3rem 0; */
 	list-style: none;
 }
 .footer_list li {
@@ -67,7 +64,6 @@ footer {
 	align-items: center;
 	justify-content: center;
 
-	/* width: 150px; */
 	width: 100%;
 	margin: 5px 10px;
 }
@@ -105,6 +101,7 @@ footer {
 
 /* media querys */
 
+/* tablet sizes */
 @media (min-width: 481px) and (max-width: 778px) {
 	footer {
 		height: 204px;
@@ -125,6 +122,7 @@ footer {
 	}
 }
 
+/* mobile sizes */
 @media (min-width: 320px) and (max-width: 480px) {
 	footer {
 		height: 232px;
