+++ src/css/thankYouPage.css
@@ -14,10 +14,10 @@
 	flex-direction: column;
 
 	width: 100%;
-	/* height: 91vh; */
 	height: 100vh;
 	margin-bottom: -1px;
 	overflow: scroll;
+
 	color: var(--white);
 	background: linear-gradient(
 			to bottom,
@@ -63,6 +63,7 @@
 
 /* media querys */
 
+/* tablet sizes */
 @media (min-width: 481px) and (max-width: 778px) {
 	.thank_section {
 		background-image: linear-gradient(
@@ -75,6 +76,7 @@
 	}
 }
 
+/* mobile sizes */
 @media (min-width: 320px) and (max-width: 480px) {
 	.thank_section {
 		background-image: linear-gradient(
