+++ src/js/index.js
@@ -1,15 +1,8 @@
-import {
-	modal,
-	navbar,
-	container_items,
-	ancor_item,
-	item_cart,
-	container_list,
-	btn_cart,
-} from "./node.js";
+import { modal } from "./node.js";
 import { viewCardSpareParts, viewCardGadgets } from "./printCard.js";
-import { viewModal, addingItem } from "./printModal.js";
-import { loadingItems, deletingItem } from "./printItemInCart.js";
+import { viewModal, addingItem, deletingItem } from "./printModal.js";
+import { loadingItems } from "./printItemInCart.js";
+import { changeColor } from "./scroll.js";
 
 window.addEventListener("DOMContentLoaded", () => {
 	// print cards within grid
@@ -20,51 +13,29 @@ window.addEventListener("DOMContentLoaded", () => {
 	loadingItems();
 });
 
+// function to open modal
 window.openModal = (id) => {
+	// print the modal
 	viewModal(id);
 };
 
+// function to close modal
 window.closeModal = () => {
 	modal.style.clipPath = "circle(0% at 50% 50%)";
 };
 
+// funcion  to add items i the cart
 window.addCart = (id) => {
 	addingItem(id);
 	closeModal();
 };
 
+// function to remove item's from the cart
 window.deleteItem = (id) => {
 	deletingItem(id);
 };
 
-// function for change the style of navbar when the user make scroll
+// function to change the style of navbar when the user scrolls
 window.onscroll = () => {
-	// onscroll permite acceder a las propiedades de la pagina al hacer scroll
-	if (window.pageYOffset > 0) {
-		// pageYOffset sera los pixeles que tenga al hacer scroll en el eje y
-		navbar.classList.add("scroll");
-		container_items.classList.add("scroll");
-		// loop for add class to ancor_item
-		// de esta forma es que se da estilo a varios elementos a la misma vez
-		for (let ancor of ancor_item) {
-			// ancor es el nombre de la variable que se usara adentro y ancor_item viene de node.js
-			ancor.classList.add("scroll");
-		}
-
-		item_cart.classList.add("scroll");
-		container_list.classList.add("scroll");
-		btn_cart.classList.add("scroll");
-	} else {
-		navbar.classList.remove("scroll");
-		container_items.classList.remove("scroll");
-		// loop for remove class to ancor_item
-		for (let ancor of ancor_item) {
-			// ancor es el nombre de la variable que se usara adentro y ancor_item viene de node.js
-			ancor.classList.remove("scroll");
-		}
-
-		item_cart.classList.remove("scroll");
-		container_list.classList.remove("scroll");
-		btn_cart.classList.remove("scroll");
-	}
+	changeColor();
 };
